requirements.txt
langgraph 
langchain_community 
langchain_openai
langchain_core
langchain-groq
python-dotenv
streamlit
ipykernel



 


